Optimize meal creation by using batch inserts for performance
